Implement dE/dh gradient calculation

+++ src/nutils.c
@@ -53,7 +53,7 @@ void hdm_vector(gsl_vector *a, gsl_vector *b, gsl_vector *r) {
 		
 void print_vector(gsl_vector *v, char *s) {
 	printf("%s", s);
-	for (int i = 0; i < v->size; i++) {
+	for (int i = 0; i < (int)v->size; i++) {
 		printf("%.2lf ", gsl_vector_get(v, i));
 	}
 	printf("\n");
@@ -61,8 +61,8 @@ void print_vector(gsl_vector *v, char *s) {
 
 void print_matrix(gsl_matrix *m, char *s) {
 	printf("%s\n", s);
-	for (int i = 0; i < m->size1; i++) {
-		for (int j = 0; j < m->size2; j++) {
+	for (int i = 0; i < (int)m->size1; i++) {
+		for (int j = 0; j < (int)m->size2; j++) {
 			printf("%.2lf ", gsl_matrix_get(m, i, j));
 		}
 		printf("\n");
